Move Home Assistant subfolder into it's own repo (#1)
These were all the changes necessary to get the Home Assistant image working after moving to it's own repo.

This also includes adding a delay to starting Home Assistant. This should allow enough time for RabbitMQ to start up. Without this, a race condition occurs and Home Assistant fails to connect to the MQTT network when both containers are started at the same time.

* Add needed configs to Dockerfile.

* Added pip command to install vscode debugging.